\name{check_valid_formula}
\alias{check_valid_formula}
\title{Check Valid Formula}
\description{Check if a molecular formula is correct or not}
\usage{
    check_valid_formula(formula = "CH4COOH")
}
\arguments{
\item{formula}{A string representation of a chemical formula.}
}
\value{
A boolean confirming the validity of a formula.
}
\examples{
    check_valid_formula("CH4COOH")
}